Fix errors, added tests

+++ app/Services/UnitQueueService.php
@@ -67,7 +67,7 @@ class UnitQueueService
         foreach ($queue_items as $item) {
             $object = $this->objects->getObjectById($item['object_id']);
 
-            $time_countdown = $item->time_end - Carbon::now()->timestamp;
+            $time_countdown = $item->time_end - (int)Carbon::now()->timestamp;
             if ($time_countdown < 0) {
                 $time_countdown = 0;
             }
@@ -80,7 +80,7 @@ class UnitQueueService
             if ($last_update < $item->time_start) {
                 $last_update = $item->time_start;
             }
-            $last_update_diff = Carbon::now()->timestamp - $last_update;
+            $last_update_diff = (int)Carbon::now()->timestamp - $last_update;
             $time_countdown_next_single = $time_per_unit - $last_update_diff;
 
             $viewModel = new UnitQueueViewModel(
@@ -187,7 +187,7 @@ class UnitQueueService
         $build_time_total = $build_time_unit * $requested_build_amount;
 
         // Time this order will start
-        $time_start = Carbon::now()->timestamp;
+        $time_start = (int)Carbon::now()->timestamp;
 
         // If there are other orders already in the queue, use the highest
         // time_end as the start time of this order.
